天哪
快点给钱吧，我真无语了，这个傻逼，什么都不懂，居然要做系统。。

+++ src/bams/dao/FeilvDAO.java
@@ -8,7 +8,7 @@ import bams.entity.Feilv;
 public interface FeilvDAO {
 
 	public List<Feilv> listAllFeilv();
-	public boolean alertFeilv(int id,double feilv);
+	public boolean alertFeilv(int id,int feilv);
 	public double getFeilv(int id);
 	public List<String> getFamilyFeilv();
 	public Connection getConnection();
